+++ girder_annotation/girder_large_image_annotation/rest/annotation.py
@@ -65,6 +65,7 @@ class AnnotationResource(Resource):
         self.route('GET', ('folder', ':id', 'present'), self.existFolderAnnotations)
         self.route('GET', ('folder', ':id', 'create'), self.canCreateFolderAnnotations)
         self.route('PUT', ('folder', ':id', 'access'), self.setFolderAnnotationAccess)
+        self.route('DELETE', ('folder', ':id'), self.deleteFolderAnnotations)
         self.route('GET', ('old',), self.getOldAnnotations)
         self.route('DELETE', ('old',), self.deleteOldAnnotations)
         self.route('GET', ('counts',), self.getItemListAnnotationCounts)
@@ -724,6 +725,27 @@ class AnnotationResource(Resource):
             count += 1
 
         return {'updated': count}
+
+    @autoDescribeRoute(
+        Description('Delete all user-owned annotations from the items in a folder')
+        .param('id', 'The ID of the folder', required=True, paramType='path')
+        .param('recurse', 'Whether or not to retrieve all '
+               'annotations from subfolders', required=False, default=False, dataType='boolean')
+        .errorResponse('ID was invalid.')
+    )
+    @access.user
+    def deleteFolderAnnotations(self, id, params):
+        setResponseTimeLimit(86400)
+        user = self.getCurrentUser()
+        if not user:
+            return []
+        count = 0
+        for annotation in self.getFolderAnnotations(id, params['recurse'], user):
+            annot = Annotation().load(annotation['_id'], user=user, getElements=False)
+            Annotation().remove(annot)
+            count += 1
+
+        return {'deleted': count}
 
     @autoDescribeRoute(
         Description('Report on old annotations.')